Support for React 18
This change adds support for React v18+ to the library. I believe many users (like myself) frequently use this lib and have encountered the same issues when running basically any npm commands that requires the use of the --force flag when using this library. This update should help resolve those issues and make it more convenient for everyone.